# bpo-39102: Increase Enum performance up to 10x times (3x average)