… Lib/test/test_dynamicclassattribute.py
… Lib/test/test_enum.py
… Misc/NEWS.d/next/Library/2019-12-20-12-31-41.bpo-39102.zqgDii.rst