# Increase Enum performance 
https://bugs.python.org/issue39102

Partial benchmark result, to see more, check [full benchmark result and source code](https://gist.github.com/Bobronium/94b5ca3e4098669d93a6127e70359307)
```py
ATTR_ACCESS:
Testing with 10000 repeats, result is average of 10 tests:

    >>> NewEnum.foo  # 0.6069349023270748 ms
    <NewEnum.foo: 1>

    >>> OldEnum.foo  # 4.5089948174334005 ms
    <OldEnum.foo: 1>

    >>> NewEnum.foo.value  # 0.965516420197984 ms
    1

    >>> OldEnum.foo.value  # 5.234090615261106 ms
    1

    >>> NewEnum.value.value  # 0.6496817059283957 ms
    3

    >>> OldEnum.value.value  # 19.718928336346387 ms
    3

    >>> try:     NewEnum.value.value = 'new' except: pass  # 4.842046525117963 ms
    AttributeError("Can't set attribute")

    >>> try:     OldEnum.value.value = 'new' except: pass  # 24.484108522222897 ms
    AttributeError("can't set attribute")


NewEnum: total: 7.0641796 ms, average: 1.1652198 ms (Fastest)
OldEnum: total: 53.9461223 ms, average: 10.3317085 ms, ~ x8.87 times slower than NewEnum 
```
## [Full benchmark result and source code](https://gist.github.com/Bobronium/94b5ca3e4098669d93a6127e70359307)

## Changes
- Optimized `Enum.__new__`
- Remove `EnumMeta.__getattr__`,
- Store `Enum.name` and `.value` in members `__dict__` for faster access
- Deprecate getting values from `Enum._name_` and `._value_`, should use public `.name` and `.value` now
- Replace `Enum._member_names_` with `._unique_member_map_` for faster lookups and iteration
- Replace `_EmumMeta._member_names` and `._last_values` with `.members` mapping (deprecation warning on using old attrs)
- Add `_str_`, `_repr_` and `_invert_` attrs to cache results of functions of the same dunder names (x35 speed boost) (948d3de)
- Various minor improvements
- Add support for direct setting and getting class attrs on `DynamicClassAttribute` without need to use slow `__getattr__`


<!-- issue-number: [bpo-39102](https://bugs.python.org/issue39102) -->
https://bugs.python.org/issue39102
<!-- /issue-number -->