@MrMrRobat 2019-12-23
> * `name` and `value` must show up in a member's `dir()`;

Yeah, looks like I just deleted them by mistake. Fixed yesterday, pushed now :)

> * `_leading_underscore_names` and `plain_names` must not be used in `EnumMeta` nor `Enum` as then the user cannot use those same names -- that's why `_sunder_` and `__dunder__` names are used instead;

Oh, I get it. So I guess it will be fine to return `list(_unique_member_map_)` or rename `_member_map` to `__member_map__`? 
First one is obviously slower, second one doesn't require any additional names. Since `_member_names_` considered deprecated and not used internally, the first one seems ok for me. What do you think?

> I do have some private code that I haven't merged in yet that does away with DynamicClassAttribute -- I'll try to get that done in the next few weeks. I think a lot of the work you have done here will still be applicable.

Does it mean that this PR won't be merged until your work is done?