@TylerYep 2020-07-31
Not a core team member, but I would really love this performance boost - any chance this will be reviewed soon?