@ethanfurman 2020-07-31
I have some other changes planned, which include some speed ups.  We'll see the changes in 3.10 (3.9 is in bug-fix only prior to first release).