@MrMrRobat 2020-08-01
> I would really love this performance boost

@TylerYep, you can check out `enum` patch I've created for current python versions, it’s a drop-in solution that enables changes from this PR along with some other improvements: https://github.com/MrMrRobat/fastenum, 
I’d appreciated any feedback on this :)

looks like I haven’t uploaded it on PyPi though. Maybe I’ll upload it on this weekend, but for now it’s still should be installable via pip:
```bash
pip install git+https://github.com/MrMrRobat/fastenum.git
```