@belm0 2020-11-23
This PR takes the approach of caching things, perhaps that is best.

However, when I looked at the Flag code several months ago (being unaware of this PR), I kept wondering **why carry such a heavy and complicated `_decompose()` function in the first place?**  A flag is just bits, right?  So I sought to eliminate it.

Proof of concept is here:  https://github.com/belm0/cpython/pull/1/files

However, this refactor is thwarted by some odd edge behavior of `Flag` and `IntFlag`, which in my opinion no one should be relying on.  To proceed with it, there would have to be will among maintainers to fix/deprecate these things:
  * repr of a compound value contains duplicate compound values (e.g. given `{A=1, B=2, C=4, AB=A|B}`, `repr(A|B|C)` is `C|AB|B|A`)
  * allowing incoherent enums which don't decompose into individual bits (e.g. `{B=3, C=4}`)
  * `IntFlag` invert behavior, which has a one-to-many mapping into the negative value space (e.g. given `{A=1, B=2}`, `~B` is `<Bar.A: -3>` rather than `<Bar.A: 1>` (and furthermore `~B == A` is `False`)