@ethanfurman 2020-11-23
@belm0 Thanks for the concise summary.

1 is a bug, and needs to be changed.
2 needs to stay.
3 needs both -- a `Flag` should behave as you say, but `IntFlag`, because it is based on integers, needs to behave like an integer would.