@belm0 2020-11-23
> 3 needs both -- a `Flag` should behave as you say, but `IntFlag`, because it is based on integers, needs to behave like an integer would.

It was a design choice, wasn't it?  `IntFlag` derives from `int` and `Flag`.  When implementing invert, it was decided to make it act like `int`.  It doesn't need to be that way, and I suspect most use cases would rather have the `Flag` behavior.

From my understanding, the primary motivation of`IntFlag` was to have a Flag that could be passed into legacy or 3rd party API.  Int-ness should have been secondary to Flag-ness.  If there is some odd API that takes such a value and does operations other than `&, |, ^`, or expects`~`to act as a signed invert, go ahead and cast to `int()` before passing it in.