@ethanfurman 2020-11-23
@belm0 If a concrete data type is mixed into an `Enum` (such as `str`, `int`, `CustomDataType`, etc.) that data type is the most important thing about it.  It would be more accurate to say "if an `Enum` is mixed into a concrete data type".

The primary motivation behind `IntFlag` was to be able to add `Enum` representation to "magic ints" that are already being used as flags -- so `IntFlag` behavior has to exactly match (or as nearly as possible) the behavior that an `int` would show.

What would be most helpful at this point is some actual code samples of "magic" integer flags that have inversion applied to them, and looking at how the surrounding code is treating the result of that operation.

Hmmm.... - just thinking out loud here, but it seems to me that the problem is that integer flags are (usually? always?) unsigned, but Python doesn't have that particular concept.  Perhaps the thing to do is to make `Flag` reveal it's `int` roots (so they could be used as indices, in calculations, etc.) but have it represent the unsigned version, and `IntFlag` could then remain the signed version.

Thoughts?