@belm0 2020-11-24
> @belm0 If a concrete data type is mixed into an `Enum` (such as `str`, `int`, `CustomDataType`, etc.) that data type is the most important thing about it. It would be more accurate to say "if an `Enum` is mixed into a concrete data type".
> 
> The primary motivation behind `IntFlag` was to be able to add `Enum` representation to "magic ints" that are already being used as flags -- so `IntFlag` behavior has to exactly match (or as nearly as possible) the behavior that an `int` would show.

I don't agree with these rationales-- it probably won't be productive for us to argue it now.

> What would be most helpful at this point is some actual code samples of "magic" integer flags that have inversion applied to them, and looking at how the surrounding code is treating the result of that operation.
> 
> Hmmm.... - just thinking out loud here, but it seems to me that the problem is that integer flags are (usually? always?) unsigned, but Python doesn't have that particular concept. Perhaps the thing to do is to make `Flag` reveal it's `int` roots (so they could be used as indices, in calculations, etc.) but have it represent the unsigned version, and `IntFlag` could then remain the signed version.

The only reasonable way to do flag negation with signed integers is to negate and then mask with the valid bits.  This is effectively what Flag does, and it's what IntFlag should do as well.   Either no one is using `IntFlag` invert, or if they are they're having to do such masking manually.

No one wants or is expecting a flag/enum component to have multiple representations (`<Bar.A: 1>`, `<Bar.A: -3>`, `<Bar.A: -5>`, ...).