@ethanfurman 2020-11-25
@belm0 : As I alluded to earlier, `IntEnum` needs to be a drop-in replacement for currently in-use magic ints -- and all the locations I looked at in the stdlib do indeed mask off the negative bits after an inversion.

As you say, it is the only reasonable way to use negated flags.

I'm more comfortable making that change now, thank you for your feedback.